chore(lint): lint css files too
