+++ .vscode/settings.json
@@ -3,7 +3,7 @@
   "typescript.format.semicolons": "remove",
   "javascript.format.semicolons": "remove",
   "typescript.suggest.includeAutomaticOptionalChainCompletions": true,
-  "editor.formatOnSave": true,
+  "editor.formatOnSave": false,
   "eslint.codeActionsOnSave.rules": [],
   "files.exclude": {
     "package-lock.json": true
@@ -12,8 +12,14 @@
   "standard.enableGlobally": false,
   "standard.run": "onSave",
   "standard.autoFixOnSave": true,
+  "editor.defaultFormatter": "standard.vscode-standard",
   "[html]": {
     "editor.defaultFormatter": "standard.vscode-standard",
-    "editor.formatOnSave": false
+  },
+  "[json]": {
+    "editor.defaultFormatter": "standard.vscode-standard",
+  },
+  "[javascript]": {
+    "editor.defaultFormatter": "standard.vscode-standard",
   }
 }
